+++ trying_polynomial_regression.cpp
@@ -8,7 +8,7 @@ int main()
     PlynomialRegression P(2);
     CSVReader csv = CSVReader("./datasets/part_2/a/2_a_dataset_1.csv");
     P.fit(csv.data[0], csv.data[1]);
-    vector<float> y = P.predict(0, 10, 0.1);
+    vector<double> y = P.predict(0, 10, 0.1);
     for (int k = 0; k < 100; k++)
     {
         cout << y[k] << endl;
